refactor : 응집도를 높이기 위한 패키지 구조 변경
