+++ src/main/java/org/programmers/springbootbasic/domain/console/Input.java
@@ -1,7 +1,7 @@
-package org.programmers.springbootbasic.io;
+package org.programmers.springbootbasic.domain.console;
 
 import org.programmers.springbootbasic.data.VoucherMainMenuCommand;
-import org.programmers.springbootbasic.dto.VoucherInputDto;
+import org.programmers.springbootbasic.domain.voucher.dto.VoucherInputDto;
 
 public interface Input {
     VoucherMainMenuCommand getVoucherMainMenuInput();
