+++ src/main/java/org/programmers/springbootbasic/domain/voucher/model/PercentDiscountVoucher.java
@@ -1,19 +1,16 @@
-package org.programmers.springbootbasic.domain;
+package org.programmers.springbootbasic.domain.voucher.model;
 
-import org.programmers.springbootbasic.exception.MinusDiscountResultException;
 import org.programmers.springbootbasic.exception.WrongRangeInputException;
 
 import java.util.UUID;
 
-public class FixedAmountVoucher implements Voucher {
-
-    private static final int MAX_FIXED_AMOUNT = Integer.MAX_VALUE;
-    private static final int MIN_FIXED_AMOUNT = 0;
-
+public class PercentDiscountVoucher implements Voucher {
+    private static final int MAX_PERCENT_AMOUNT = 100;
+    private static final int MIN_PERCENT_AMOUNT = 0;
     private final UUID voucherId;
     private final long amount;
 
-    public FixedAmountVoucher(UUID voucherId, long amount) {
+    public PercentDiscountVoucher(UUID voucherId, long amount) {
         this.voucherId = voucherId;
         this.amount = amount;
         validateAmountRange();
@@ -31,8 +28,8 @@ public class FixedAmountVoucher implements Voucher {
 
     @Override
     public long discount(long originalPrice) {
-        if (originalPrice - amount < 0) throw new MinusDiscountResultException("할인된 금액은 0원보다 작을 수 없습니다.");
-        return originalPrice - amount;
+        // 실세계에서 할인 시 소수 단위는 제거하므로 long type으로 정의
+        return (long) (originalPrice - ((double) amount / (double) 100));
     }
 
     @Override
@@ -43,7 +40,7 @@ public class FixedAmountVoucher implements Voucher {
     }
 
     private void validateAmountRange() {
-        if (!(MIN_FIXED_AMOUNT <= amount && amount <= MAX_FIXED_AMOUNT))
+        if (!(MIN_PERCENT_AMOUNT <= amount && amount <= MAX_PERCENT_AMOUNT))
             throw new WrongRangeInputException("올바르지 않은 Amount의 값 범위입니다.");
     }
 }
