--- src/main/java/org/programmers/springbootbasic/repository/VoucherRepository.java
@@ -1,11 +1,0 @@
-package org.programmers.springbootbasic.repository;
-
-import org.programmers.springbootbasic.domain.Voucher;
-
-import java.util.List;
-
-public interface VoucherRepository {
-    Voucher save(Voucher voucher);
-
-    List<Voucher> findAll();
-}
